requirements.txt
PIllow
PyAutoGUI
win10toast